code refactored
stages are now a caallback of current stage

the color calculation is inside of a function now, todo to LIMIT color and perhaps actually set the text drawing inside of a STAGE code

the switch statement for  argument is odd but works, i guess, plus it's fun, although i don't know how it is like on windows and other cpu architectures. hope its fine... if not... RIP.

added velocity to objects. hey, i forgot about idea of adding some parameter to them or optimizing these... heh... dammit.

also, SceneManager added as a somewhat structured way to add and remove objects in the future, also really used as a way for me to rethink my code

reminder for myself in future: i tihnk constexpr in a while loop is not looking good

reminder to myself - i came back with thought of writing objects in from a .json file, but in the end i didnt do it yet...

also thought about idea of animating, means i'll have to think about timing things... and i also thought about objects that could make dialogues... sounds like a tough thing, may have to consider a way to turn it into a separate file or make short as hell. thinking RN i remembered that i wanted to have a function pointer for any object in array, that can be used as a function ON INTERACTION!